chore(4615): extract editor context and related injection utils into a shared editor module
